requirements.txt
mysql_connector_repackaged==0.3.1
PyQt5==5.15.1
